requirements.txt
supabase==2.12.0
python-dotenv==1.0.1
postgrest==0.16.2